fix: resolve compilation errors after Arsenal refactoring
Fix multiple compilation errors introduced during the refactoring process:

API Interface Fixes:
- Fix ArsenalAppBar parameter requirements (title, searchField, isSearching, onBack, onToggleSearch)
- Fix ArsenalListView extraInfoBuilder required parameter
- Fix BagInfo property access (use 'name' instead of 'displayName')

Type Safety Improvements:
- Fix dynamic to int cast in supabase_user_arsenal_repository.dart
- Add missing BuildContext parameter to _buildBagOptionTile method
- Add missing UserProfile import to arsenal_bag_service.dart

Provider and Service Updates:
- Convert ArsenalDialogService from provider to static class
- Fix BallFilters property access (use 'brands', 'cores', 'coverstocks' instead of 'selectedBrands', 'selectedCategories')
- Update UserArsenalInstance property references (use 'belongsToCategory' method instead of 'category' property)

Filter System Adjustments:
- Remove unsupported weight and games used filters from BallFilters integration
- Update filter summary and validation logic to match BallFilters schema
- Use brandName as fallback for category operations where needed

These fixes ensure the refactored Arsenal feature compiles correctly while maintaining architectural improvements from the 4-phase refactoring process.

+++ lib/features/arsenal/data/repositories/supabase_user_arsenal_repository.dart
@@ -74,7 +74,7 @@ class SupabaseUserArsenalRepository implements UserArsenalRepository {
 
 
       // Fetch the complete instance with bowling ball data
-      return _getCompleteInstance(response['id']);
+      return _getCompleteInstance(response['id'] as int);
     } catch (e) {
       throw Exception('Failed to add ball from library: $e');
     }
